+++ crypto/src/commonMain/kotlin/at/asitplus/awesn1/crypto/pki/Pkcs10CertificationRequestInfo.kt
@@ -5,6 +5,7 @@ package at.asitplus.awesn1.crypto.pki
 
 import at.asitplus.awesn1.Asn1Exception
 import at.asitplus.awesn1.crypto.SubjectPublicKeyInfo
+import at.asitplus.awesn1.serialization.LenientSet
 import at.asitplus.awesn1.serialization.Asn1Tag
 import kotlinx.serialization.Serializable
 import kotlin.experimental.ExperimentalObjCRefinement
@@ -32,18 +33,19 @@ import kotlin.native.HiddenFromObjC
  * }
  * ```
  *
- * [rawAttributes] uses [Pkcs10CsrAttributeBag] to hold malformed data too. **DO NOT ASSUME ANY PARTICULAR ORDER OR CONCRETE COLLECTION TYPE!** Canonical sorting happens only on encode!
+ * [rawAttributes] uses [LenientSet] to hold malformed data too. **DO NOT ASSUME ANY PARTICULAR ORDER OR CONCRETE COLLECTION TYPE!** Canonical sorting happens only on encode!
  * Hence, the order of attributes may differ post-encode.
  *
  * This class's [equals] and [hashCode] reflect this characteristic: Order of attributes is irrelevant for equality!
  */
+@ConsistentCopyVisibility
 @Serializable
 data class Pkcs10CertificationRequestInfo private constructor(
     val version: Version = Version.V1,
     val subjectName: X500Name,
     val publicKey: SubjectPublicKeyInfo,
     @Asn1Tag(tagNumber = 0u)
-    val rawAttributes: Pkcs10CsrAttributeBag = Pkcs10CsrAttributeBag(emptySet()),
+    val rawAttributes: LenientSet<Pkcs10CsrAttribute> = LenientSet(),
 ) {
 
     constructor(
@@ -51,7 +53,11 @@ data class Pkcs10CertificationRequestInfo private constructor(
         subjectName: X500Name,
         publicKey: SubjectPublicKeyInfo,
         attributes: Set<Pkcs10CsrAttribute> = emptySet(),
-    ) : this(version, subjectName, publicKey, rawAttributes = Pkcs10CsrAttributeBag(attributes))
+    ) : this(version, subjectName, publicKey, rawAttributes = LenientSet(attributes)) {
+        require(attributes.distinctBy { it.oid }.size == attributes.size) {
+            "Multiple CSR attributes with the same OID found"
+        }
+    }
 
 
     /**
@@ -61,13 +67,14 @@ data class Pkcs10CertificationRequestInfo private constructor(
      */
     @OptIn(ExperimentalObjCRefinement::class)
     @Suppress("WRONG_ANNOTATION_TARGET_WITH_USE_SITE_TARGET")
-    @get:Throws(IllegalArgumentException::class)
+    @get:Throws(Asn1Exception::class)
     @HiddenFromObjC
     @get:HiddenFromObjC
     val attributes: Set<Pkcs10CsrAttribute>
-        get() = if (rawAttributes.distinctBy { it.oid }.size != rawAttributes.size)
-            throw Asn1Exception("Multiple CSR attributes with the same OID found")
-        else rawAttributes.toSet()
+        get() = rawAttributes.toValidatedSet().also {
+            if (it.distinctBy { attribute -> attribute.oid }.size != it.size)
+                throw Asn1Exception("Multiple CSR attributes with the same OID found")
+        }
 
 
     /**